+++ packages/blade-core/src/utils/index.ts
@@ -27,3 +27,4 @@ export * from './styledProps';
 export * from './tokenToCSSVariable';
 export * from './types';
 export * from './amount';
+export * from './useId';